TeleService: Prompt user the fixed dialing number can’t be empty
It does not check if the fixed dialing number is null or not,
so it also can be saved as FDN when the fixed dialing number is null.

When the number is null, we do not save it as FDN, and show a toast
to prompt user the fixed dialing number can’t be empty.

CRs-Fixed: 522515

Change-Id: I0a78be0657cdd35ed00058bb825d2e077f52b471